Polishing - Move Test Cases
